refactor(undo): rename undo/undo.py to undo/core.py
