+++ src/griptape_nodes/retained_mode/managers/undo/core.py
